Pdfifier/requirements.txt
# Add your dependencies here
fastapi
uvicorn
requests
beautifulsoup4
fpdf
python-dotenv
